Refactored parameter substitution
Making it more universally accessible for several places that need to display requirement text
