+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Business/Question/QuestionBusiness.cs
@@ -5,7 +5,6 @@
 // 
 //////////////////////////////// 
 using CSETWebCore.Business.Standards;
-using CSETWebCore.DataLayer.Manual;
 using CSETWebCore.DataLayer.Model;
 using CSETWebCore.Interfaces.Common;
 using CSETWebCore.Interfaces.Document;
@@ -14,13 +13,8 @@ using CSETWebCore.Interfaces.Question;
 using CSETWebCore.Model.Hydro;
 using CSETWebCore.Model.Question;
 using Nelibur.ObjectMapper;
-using Npoi.Mapper;
-using NPOI.HPSF;
-using NPOI.SS.UserModel;
-using NPOI.Util;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 
 namespace CSETWebCore.Business.Question
@@ -37,6 +31,7 @@ namespace CSETWebCore.Business.Question
         private readonly IQuestionRequirementManager _questionRequirement;
         private readonly IAssessmentUtil _assessmentUtil;
         private readonly CSETContext _context;
+
 
         /// <summary>
         /// Constructor.
@@ -294,7 +289,8 @@ namespace CSETWebCore.Business.Question
             var qvm = new QuestionDetailsBusiness(
                 new StandardSpecficLevelRepository(_context),
                 new InformationTabBuilder(_context, _htmlConverter, _tokenManager),
-                _context, _tokenManager, _document
+                _context, _tokenManager, _document, 
+                _assessmentUtil, _questionRequirement
             );
 
             _questionRequirement.InitializeManager(_tokenManager.AssessmentForUser());
