+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Business/Question/RequirementBusiness.cs
@@ -27,6 +27,7 @@ namespace CSETWebCore.Business.Question
 
         private readonly IAssessmentUtil _assessmentUtil;
         private readonly IQuestionRequirementManager _questionRequirement;
+        private readonly ParameterSubstitution _parmSub;
         private readonly ITokenManager _tokenManager;
         private CSETContext _context;
 
@@ -41,6 +42,8 @@ namespace CSETWebCore.Business.Question
             _questionRequirement = questionRequirement;
             _tokenManager = tokenManager;
             _context = context;
+
+            _parmSub = new ParameterSubstitution(context, tokenManager);
         }
 
         public void SetRequirementAssessmentId(int assessmentId)
@@ -235,7 +238,7 @@ namespace CSETWebCore.Business.Question
                     qa.HasObservation = answer.b.HasDiscovery ?? false;
                 }
 
-                qa.ParmSubs = GetTokensForRequirement(qa.QuestionId, (answer != null) ? answer.a.Answer_Id : 0);
+                qa.ParmSubs = _parmSub.GetTokensForRequirement(qa.QuestionId, (answer != null) ? answer.a.Answer_Id : 0);
 
                 subcat.Questions.Add(qa);
             }
@@ -367,134 +370,6 @@ namespace CSETWebCore.Business.Question
 
 
         /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="db"></param>
-        public void LoadParametersList()
-        {
-            SetRequirementAssessmentId(_tokenManager.AssessmentForUser());
-
-            parametersDictionary = (from p in _context.PARAMETERS
-                                    join r in _context.PARAMETER_REQUIREMENTS on p.Parameter_ID equals r.Parameter_Id
-                                    select new { p, r }).AsEnumerable()
-                .GroupBy(x => x.r.Requirement_Id)
-                .ToDictionary(x => x.Key, x => x.Select(y => y.p).ToList());
-
-
-            parametersAssessmentList = (from pa in _context.PARAMETER_ASSESSMENT
-                                        join p in _context.PARAMETERS on pa.Parameter_ID equals p.Parameter_ID
-                                        where pa.Assessment_ID == _questionRequirement.AssessmentId
-                                        select new ParameterAssessment() { p = p, pa = pa }).ToList();
-
-            parametersAnswerDictionary = (from p in _context.PARAMETERS
-                                          join pv in _context.PARAMETER_VALUES on p.Parameter_ID equals pv.Parameter_Id
-                                          select new ParameterValues() { p = p, pv = pv }).AsEnumerable()
-            .GroupBy(x => x.pv.Answer_Id)
-            .ToDictionary(x => x.Key, x => x.Select(y => y).ToList());
-        }
-
-        private Dictionary<int, List<PARAMETERS>> parametersDictionary = null;
-        private List<ParameterAssessment> parametersAssessmentList;
-        private Dictionary<int, List<ParameterValues>> parametersAnswerDictionary;
-
-
-        /// <summary>
-        /// Pull any 'global' parameters for the requirement, overlaid with any 'local' parameters for the answer.
-        /// </summary>
-        /// <param name="reqId"></param>
-        /// <param name="ansId"></param>
-        /// <returns></returns>
-        public List<ParameterToken> GetTokensForRequirement(int reqId, int ansId)
-        {
-            ParameterSubstitution ps = new ParameterSubstitution();
-
-            // get the 'base' parameter values (parameter_name) for the requirement
-            if (parametersDictionary == null)
-            {
-                LoadParametersList();
-            }
-
-            List<PARAMETERS> qBaseLevel;
-            if (parametersDictionary.TryGetValue(reqId, out qBaseLevel))
-            {
-                foreach (var b in qBaseLevel)
-                {
-                    ps.Set(b.Parameter_ID, b.Parameter_Name, null, reqId, 0);
-                }
-            }
-
-            // overlay with any assessment-specific parameters for the requirement
-            var qAssessLevel = parametersAssessmentList;
-
-            foreach (var a in qAssessLevel)
-            {
-                ps.Set(a.p.Parameter_ID, a.p.Parameter_Name, a.pa.Parameter_Value_Assessment, reqId, 0);
-            }
-
-            // overlay with any 'inline' values for the answer
-            if (ansId != 0)
-            {
-                List<ParameterValues> qLocal;
-                if (parametersAnswerDictionary.TryGetValue(ansId, out qLocal))
-                {
-                    foreach (var local in qLocal.ToList())
-                    {
-                        ps.Set(local.p.Parameter_ID, local.p.Parameter_Name, local.pv.Parameter_Value, reqId, local.pv.Answer_Id);
-                    }
-                }
-            }
-
-            ps.Tokens = ps.Tokens.OrderByDescending(x => x.Token.Length).ToList();
-
-            return ps.Tokens;
-        }
-
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public List<ParameterToken> GetDefaultParametersForAssessment()
-        {
-            SetRequirementAssessmentId(_tokenManager.AssessmentForUser());
-
-            ParameterSubstitution ps = new ParameterSubstitution();
-
-            // Get the list of requirement IDs
-            List<RequirementPlus> reqs = GetControls().Requirements.ToList();
-            List<int> requirementIds = reqs.Select(r => r.Requirement.Requirement_Id).ToList();
-
-
-            // get the 'base' parameter values (parameter_name) for the requirement
-            var qBaseLevel = from p in _context.PARAMETERS
-                             join r in _context.PARAMETER_REQUIREMENTS on p.Parameter_ID equals r.Parameter_Id
-                             where requirementIds.Contains(r.Requirement_Id)
-                             select new { p, r };
-
-            foreach (var b in qBaseLevel)
-            {
-                ps.Set(b.p.Parameter_ID, b.p.Parameter_Name, b.p.Parameter_Name, b.r.Requirement_Id, 0);
-            }
-
-            // overlay with any assessment-specific parameters for the requirement
-            var qAssessLevel = from pa in _context.PARAMETER_ASSESSMENT
-                               join p in _context.PARAMETERS on pa.Parameter_ID equals p.Parameter_ID
-                               join pr in _context.PARAMETER_REQUIREMENTS on p.Parameter_ID equals pr.Parameter_Id
-                               where pa.Assessment_ID == _questionRequirement.AssessmentId
-                                && requirementIds.Contains(pr.Requirement_Id)
-                               select new { p, pa, pr };
-
-            foreach (var a in qAssessLevel)
-            {
-                ps.Set(a.p.Parameter_ID, a.p.Parameter_Name, a.pa.Parameter_Value_Assessment, a.pr.Requirement_Id, 0);
-            }
-
-
-            return ps.Tokens;
-        }
-
-
-        /// <summary>
         /// Saves a new text value override in a PARAMETER_ASSESSMENT row.  
         /// </summary>
         /// <param name="requirementId"></param>
@@ -552,123 +427,6 @@ namespace CSETWebCore.Business.Question
 
             // build a partial return object just to inform the UI what the new value is
             return new ParameterToken(pa.Parameter_ID, "", pa.Parameter_Value_Assessment, 0, 0);
-        }
-
-
-        /// <summary>
-        /// Saves a new text value override in a PARAMETER_VALUES row.  
-        /// Creates a new Answer if need be.
-        /// If no text is provided, any PARAMETER_VALUES is deleted.
-        /// </summary>
-        /// <param name="requirementId"></param>
-        /// <param name="parameterId"></param>
-        /// <param name="answerId"></param>
-        /// <param name="newText"></param>
-        /// <returns></returns>
-        public ParameterToken SaveAnswerParameter(int requirementId, int parameterId, int answerId, string newText)
-        {
-            SetRequirementAssessmentId(_tokenManager.AssessmentForUser());
-
-            // create an answer if there isn't one already
-            if (answerId == 0)
-            {
-                Answer ans = new Answer()
-                {
-                    QuestionId = requirementId,
-                    QuestionType = "Requirement",
-                    MarkForReview = false,
-                    QuestionNumber = "0",
-                    AnswerText = "U"
-                };
-                answerId = _questionRequirement.StoreAnswer(ans);
-            }
-
-            // If an empty value is supplied, delete the PARAMETER_VALUES row.
-            if (string.IsNullOrEmpty(newText))
-            {
-                var g = _context.PARAMETER_VALUES.Where(pv => pv.Parameter_Id == parameterId && pv.Answer_Id == answerId).FirstOrDefault();
-                if (g != null)
-                {
-                    _context.PARAMETER_VALUES.Remove(g);
-                    _context.SaveChanges();
-                }
-
-                _assessmentUtil.TouchAssessment(_questionRequirement.AssessmentId);
-
-                return this.GetTokensForRequirement(requirementId, answerId).Where(p => p.Id == parameterId).First();
-            }
-
-
-            // Otherwise, add or update the PARAMETER_VALUES row
-            var dbParameterValues = _context.PARAMETER_VALUES.Where(pv => pv.Parameter_Id == parameterId
-                    && pv.Answer_Id == answerId).FirstOrDefault();
-
-            if (dbParameterValues == null)
-            {
-                dbParameterValues = new PARAMETER_VALUES();
-            }
-
-            dbParameterValues.Answer_Id = answerId;
-            dbParameterValues.Parameter_Id = parameterId;
-            dbParameterValues.Parameter_Is_Default = false;
-            dbParameterValues.Parameter_Value = newText;
-
-
-            if (_context.PARAMETER_VALUES.Find(dbParameterValues.Answer_Id, dbParameterValues.Parameter_Id) == null)
-            {
-                _context.PARAMETER_VALUES.Add(dbParameterValues);
-            }
-            else
-            {
-                _context.PARAMETER_VALUES.Update(dbParameterValues);
-            }
-            _context.SaveChanges();
-
-            _assessmentUtil.TouchAssessment(_questionRequirement.AssessmentId);
-
-            // Return the token that was just updated
-            return this.GetTokensForRequirement(requirementId, answerId).Where(p => p.Id == parameterId).First();
-        }
-
-
-        /// <summary>
-        /// Returns Requirement text with Parameter substitutions applied.
-        /// Also converts linefeed characters to HTML.
-        /// </summary>
-        /// <param name="requirementText"></param>
-        /// <returns></returns>
-        public string ResolveParameters(int reqId, int ansId, string requirementText)
-        {
-            List<ParameterToken> tokens = this.GetTokensForRequirement(reqId, ansId);
-            foreach (ParameterToken t in tokens)
-            {
-                if (t.Substitution != null)
-                {
-                    requirementText = requirementText.Replace(
-                        "{{" + t.Token + "}}",
-                        "<em>" + t.Substitution + "</em>");
-                }
-            }
-
-            // format anything still unresolved
-            requirementText = requirementText.Replace("{{", "[<em>").Replace("}}", "</em>]");
-
-            return requirementText;
-        }
-
-
-        public string RichTextParameters(int reqId, int ansId, string requirementText)
-        {
-            List<ParameterToken> tokens = this.GetTokensForRequirement(reqId, ansId);
-            foreach (ParameterToken t in tokens)
-            {
-                requirementText = requirementText.Replace(t.Token, t.Substitution);
-            }
-
-            requirementText = requirementText.Replace("\r\n", "%0D%0A").Replace("\r", "%0D%0A").Replace("\n", "%0D%0A");
-
-            return requirementText;
-
         }
     }
 }
