+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Business/Reports/ReportsData/ReportsDataBusiness.cs
@@ -534,7 +534,7 @@ namespace CSETWebCore.Business.Reports
         public async Task<List<StandardQuestions>> GetStandardQuestionAnswers(int assessId)
         {
             CsetwebContextProcedures context = new CsetwebContextProcedures(_context);
-            var rm = new Question.RequirementBusiness(_assessmentUtil, _questionRequirement, _context, _tokenManager);
+            var parmSub = new ParameterSubstitution(_context, _tokenManager);
             var dblist = await context.usp_GetQuestionsAsync(assessId);
 
             List<StandardQuestions> list = new List<StandardQuestions>();
@@ -556,7 +556,7 @@ namespace CSETWebCore.Business.Reports
                 {
                     ShortName = a.ShortName,
 
-                    Question = rm.ResolveParameters(a.QuestionOrRequirementID, a.AnswerID, a.QuestionText),
+                    Question = parmSub.ResolveParameters(a.QuestionOrRequirementID, a.AnswerID, a.QuestionText),
                     QuestionId = a.QuestionOrRequirementID,
                     Answer = a.AnswerText,
                     CategoryAndNumber = a.CategoryAndNumber
@@ -636,10 +636,9 @@ namespace CSETWebCore.Business.Reports
         {
             var results = new List<QuestionsWithAltJust>();
 
-            var rm = new Question.RequirementBusiness(_assessmentUtil, _questionRequirement, _context, _tokenManager);
+            var parmSub = new ParameterSubstitution(_context, _tokenManager);
 
             // get any "A" answers that currently apply
-
             var relevantAnswers = new RelevantAnswers().GetAnswersForAssessment(_assessmentId, _context)
                 .Where(ans => ans.Answer_Text == "A").ToList();
 
@@ -669,7 +668,7 @@ namespace CSETWebCore.Business.Reports
 
                 foreach(var req in reqs)
                 {
-                    req.Question = rm.ResolveParameters(req.Id, req.AnswerId, req.Question);
+                    req.Question = parmSub.ResolveParameters(req.Id, req.AnswerId, req.Question);
                 }
 
                 return reqs;
@@ -703,7 +702,8 @@ namespace CSETWebCore.Business.Reports
         {
             var results = new List<QuestionsWithComments>();
 
-            var rm = new Question.RequirementBusiness(_assessmentUtil, _questionRequirement, _context, _tokenManager);
+            var parmSub = new ParameterSubstitution(_context, _tokenManager);
+
             // get any "marked for review" or commented answers that currently apply
             var relevantAnswers = new RelevantAnswers().GetAnswersForAssessment(_assessmentId, _context)
                 .Where(ans => !string.IsNullOrEmpty(ans.Comment))
@@ -725,7 +725,7 @@ namespace CSETWebCore.Business.Reports
                             {
                                 Answer = ans.Answer_Text,
                                 CategoryAndNumber = req.Standard_Category + " - " + req.Requirement_Title,
-                                Question = rm.ResolveParameters(ans.Question_Or_Requirement_ID, ans.Answer_ID, req.Requirement_Text),
+                                Question = parmSub.ResolveParameters(ans.Question_Or_Requirement_ID, ans.Answer_ID, req.Requirement_Text),
                                 Comment = ans.Comment
                             };
 
@@ -756,7 +756,7 @@ namespace CSETWebCore.Business.Reports
         /// <returns></returns>
         public List<QuestionsMarkedForReview> GetQuestionsMarkedForReview()
         {
-            var rm = new Question.RequirementBusiness(_assessmentUtil, _questionRequirement, _context, _tokenManager);
+            var parmSub = new ParameterSubstitution(_context, _tokenManager);
 
             var results = new List<QuestionsMarkedForReview>();
 
@@ -790,7 +790,7 @@ namespace CSETWebCore.Business.Reports
 
                 foreach (var req in reqs)
                 {
-                    req.Question = rm.ResolveParameters(req.Id, req.AnswerId, req.Question);
+                    req.Question = parmSub.ResolveParameters(req.Id, req.AnswerId, req.Question);
                 }
 
                 return reqs;
@@ -820,7 +820,8 @@ namespace CSETWebCore.Business.Reports
         public List<QuestionsMarkedForReview> GetQuestionsReviewed()
         {
             var results = new List<QuestionsMarkedForReview>();
-            var rm = new Question.RequirementBusiness(_assessmentUtil, _questionRequirement, _context, _tokenManager);
+
+            var parmSub = new ParameterSubstitution(_context, _tokenManager);
 
             // get any "marked for review" or commented answers that currently apply
             var relevantAnswers = new RelevantAnswers().GetAnswersForAssessment(_assessmentId, _context)
@@ -845,14 +846,14 @@ namespace CSETWebCore.Business.Reports
                                 Answer = ans.Answer_Text,
                                 AnswerId = ans.Answer_ID,
                                 CategoryAndNumber = req.Standard_Category + " - " + req.Requirement_Title,
-                                Question = rm.ResolveParameters(ans.Question_Or_Requirement_ID, ans.Answer_ID, req.Requirement_Text)
+                                Question = parmSub.ResolveParameters(ans.Question_Or_Requirement_ID, ans.Answer_ID, req.Requirement_Text)
                             };
 
                 var reqs = query.ToList();
 
                 foreach (var req in reqs)
                 {
-                    req.Question = rm.ResolveParameters(req.Id, req.AnswerId, req.Question);
+                    req.Question = parmSub.ResolveParameters(req.Id, req.AnswerId, req.Question);
                 }
 
                 return reqs;
@@ -879,7 +880,7 @@ namespace CSETWebCore.Business.Reports
         {
             var lang = _tokenManager.GetCurrentLanguage();
 
-            var rm = new Question.RequirementBusiness(_assessmentUtil, _questionRequirement, _context, _tokenManager);
+            var parmSub = new ParameterSubstitution(_context, _tokenManager);
 
             List<RankedQuestions> list = new List<RankedQuestions>();
             List<usp_GetRankedQuestions_Result> rankedQuestionList = _context.usp_GetRankedQuestions(_assessmentId).ToList();
@@ -894,8 +895,7 @@ namespace CSETWebCore.Business.Reports
                     }
                 }
 
-
-                q.QuestionText = rm.ResolveParameters(q.QuestionOrRequirementID, q.AnswerID, q.QuestionText);
+                q.QuestionText = parmSub.ResolveParameters(q.QuestionOrRequirementID, q.AnswerID, q.QuestionText);
 
                 q.Category = _overlay.GetPropertyValue("STANDARD_CATEGORY", q.Category.ToLower(), lang) ?? q.Category;
 
@@ -912,11 +912,12 @@ namespace CSETWebCore.Business.Reports
             return list;
         }
 
+
         public List<PhysicalQuestions> GetQuestionsWithSupplementals()
         {
             var lang = _tokenManager.GetCurrentLanguage();
 
-            var rm = new Question.RequirementBusiness(_assessmentUtil, _questionRequirement, _context, _tokenManager);
+            var parmSub = new ParameterSubstitution(_context, _tokenManager);
 
             List<PhysicalQuestions> list = new List<PhysicalQuestions>();
             List<usp_GetRankedQuestions_Result> rankedQuestionList = _context.usp_GetRankedQuestions(_assessmentId).ToList();
@@ -940,7 +941,7 @@ namespace CSETWebCore.Business.Reports
                 }
 
 
-                q.QuestionText = rm.ResolveParameters(q.QuestionOrRequirementID, q.AnswerID, q.QuestionText);
+                q.QuestionText = parmSub.ResolveParameters(q.QuestionOrRequirementID, q.AnswerID, q.QuestionText);
 
                 q.Category = _overlay.GetPropertyValue("STANDARD_CATEGORY", q.Category.ToLower(), lang) ?? q.Category;
                 var comment = _context.Answer_Requirements.Where(x => x.Question_Or_Requirement_Id == q.QuestionOrRequirementID).FirstOrDefault()?.Comment;
