+++ CSETWebNg/src/app/assessment/questions/question-block/question-block.component.html
@@ -141,7 +141,7 @@
         </div>
         <div class="w-100 d-flex flex-row" *ngIf="!layoutSvc.hp">
           <div style="width: 100%;list-style-type: none;">
-            <div [style.word-break]="applyWordBreak(q)" [innerHTML]="applyTokensToText(q)"
+            <div [style.word-break]="applyWordBreak(q)" [innerHTML]="questionsSvc.applyTokensToText(q)"
               (click)="questionTextClicked(q, $event)">
             </div>
             <span *ngIf="showQuestionIds" class="debug-highlight">{{ q.questionId }}</span>
@@ -153,7 +153,7 @@
           <div class="">
             <div style="font-weight: bold; margin-right: 1em; white-space: nowrap;">{{ q.displayNumber }}</div>
             <div class="w-100 mb-3">
-              <div [style.word-break]="applyWordBreak(q)" [innerHTML]="applyTokensToText(q)"
+              <div [style.word-break]="applyWordBreak(q)" [innerHTML]="questionsSvc.applyTokensToText(q)"
                 (click)="questionTextClicked(q, $event)">
               </div>
               <span *ngIf="showQuestionIds" class="debug-highlight">{{ q.questionId }}</span>
