+++ CSETWebNg/src/app/services/questions.service.ts
@@ -30,6 +30,7 @@ import { AssessmentService } from './assessment.service';
 import { QuestionFilterService } from './filtering/question-filter.service';
 import { BehaviorSubject } from 'rxjs';
 import { TranslocoService } from '@jsverse/transloco';
+import { LinebreakPipe } from '../helpers/linebreak.pipe';
 
 const headers = {
   headers: new HttpHeaders()
@@ -71,7 +72,8 @@ export class QuestionsService {
     private configSvc: ConfigService,
     private tSvc: TranslocoService,
     private assessmentSvc: AssessmentService,
-    private questionFilterSvc: QuestionFilterService
+    private questionFilterSvc: QuestionFilterService,
+    public linebreakPipe: LinebreakPipe
   ) { }
 
   /**
@@ -437,4 +439,62 @@ export class QuestionsService {
     return this.http.get(this.configSvc.apiUrl + 'getRegulatoryCitations?questionId=' + questionId)
   }
 
+
+  /**
+   * If there are any parameters in the text defined by double curly braces
+   * format them to look like the published standard.  No substitution is
+   * performed by this function, just formatting.
+   */
+  formatParameters(text: string) {
+    text = text.replace(/{{/g, '[<em>').replace(/}}/g, '</em>]');
+    return text;
+  }
+
+  /**
+   * Replace parameter placeholders in the question text template with any overridden values.
+   * @param q
+   */
+  applyTokensToText(q: Question) {
+    let text = q.questionText;
+
+    text = this.linebreakPipe.transform(text);
+
+    if (!q.parmSubs) {
+      return text;
+    }
+
+    // Substitute the longer tokens first, to avoid substituting a nested token
+    // and leave the outer token untouched.  We currently only support the outer tokens.
+    q.parmSubs.sort((a, b) => {
+      if (a.token.length > b.token.length) return -1;
+      if (a.token.length < b.token.length) return 1;
+      return 0;
+    });
+
+    q.parmSubs.forEach(t => {
+      if (t.substitution == null) {
+        // uncustomized
+        text = this.replaceAll(text, `{{${t.token}}}`, "[<span class='sub-me fst-italic pid-" + t.id + "'>" + t.token + "</span>]");
+      } else {
+        // customized
+        text = this.replaceAll(text, `{{${t.token}}}`, "<span class='sub-me pid-" + t.id + "'>" + t.substitution + "</span>");
+      }
+    });
+
+    return text;
+  }
+
+  /**
+   * 
+   * @param origString 
+   * @param searchStr 
+   * @param replaceStr 
+   * @returns 
+   */
+  replaceAll(origString: string, searchStr: string, replaceStr: string) {
+    // escape regexp special characters in search string
+    searchStr = searchStr.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
+
+    return origString.replace(new RegExp(searchStr, 'gi'), replaceStr);
+  }
 }
